backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sendgrid==6.10.0
schedule==1.2.0
asyncpg==0.29.0